entregador-app/requirements.txt
Werkzeug==2.3.7
Flask==2.3.3
mysql-connector-python==8.1.0
flask-restx==1.2.0
flasgger==0.9.7.1
Flask-Testing==0.8.1
